Arrows sound, paint bomb particles/sound, pillar explosion particles/sound, reorganised prefab folders, tested async/await functionality & fixed multiplier
- Arrow sounds are now assigned in the editor so the paint bomb sound
can be attached when needed.

- Paint bomb particles are loaded into the scene while scene loads and
then played when needed, help with performance.
- Paint bomb sound is attached to arrow script so its not called on
food start, performance boost.

- Pillar explosion particles loaded in during scene load and played
when needed for performance.
- Pillar sound plays for pillar when exploding for performance boost.

- Reorganised prefab folders for Pillar of Light as it was getting
quite messy.
- Moved explosion particles effects and materials into these folders
well.

- Tested and implemented the .NET 4 Await/Async functionality for
practice into the Time Controller script. Can be reverted in the future
if needed.

- Score multiplier UI was causing massive performance issues so I
removed spawning of score multiplier and replaced it with a small
particle effect that burst one particle corresponding to the current
multiplier.